Improving support for foreign customers (#17)
Export customers don't have a kennitala and may need to be dealt with very differently. This adds a mechanism to generate kennitala-like strings for DK to use as a customer number.

International orders can be taken out of the invoice automation mechanism, a separate placeholder kennitala for foreign customers is defined and a prefix for the generated customer number can be defined.