Introducing customer prices and discounts
This massive change introduces support for customer discounts and price groups from DK as well as downstream customer sync from DK for the relevant properties.

Customer prices and discounts work for logged-in customers with the kennitala attribute set to a value corresponding to the DK customer record.

This also upgrades product price sync to fetch data on price 2 and 3 from DK.

Fill support for variable products still remains a work in progress. Currently, price groups do not work with variations.