Introducing support for product discounts
Customer discounts are now indicated on invoices and in the Order Editor.

This also enables invoices to be generated, even after removing any of the products from WooCommerce.

A meta value is set for new orders, indicating the version number of the plugin, which prevents invoices from being generated if the plugin was not installed when the order took place. This is due to all the meta values that get assigned to order items and need to be in place for invoicing to happen.